Experimental rearrangement of the towers view
One of the things I liked about the old view was that it was easy to
scan down the right column to get a sense of how long various towers
had. This is an attempt to bring that back, with some of the shiny new
information display.  I'm not sure how well it works, curious for
thoughts.

If we do decide we want to go in some direction like this, it will
definitely need some polish. For example, the fact that the meters don't
line up vertically drives me nuts....